made bar chart as a reusable component
