+++ src/components/BarChart/index.js
@@ -1,14 +1,12 @@
 import { Chart } from "react-google-charts";
 import React from "react";
-import RevenueChart from '../RevenueChart/index';
 
-console.log("options",this.options);
-export default function BarChart() {
+export default function BarChart(props) {
         return (
             <Chart
                 chartType="BarChart"
-                data={this.chartData}
-                options={this.options}
+                data={props.data}
+                options={props.options}
                 width="100%"
                 height="300px"
             />
